lexi/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
kaggle==1.5.16
pytesseract==0.3.10
aiohttp==3.9.1
httpx==0.25.2
psycopg2-binary==2.9.9
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.4

# Monitoring
prometheus-client==0.19.0